requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
transformers>=4.30.0
olmocr>=0.1.76
pillow>=9.0.0
werkzeug>=2.3.0 